Split LocalizationMethod into its own source file
