+++ lib/go-tc/enum/enum.go
@@ -29,7 +29,6 @@ package enum
  */
 
 import (
-	"database/sql/driver"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -251,83 +250,6 @@ func (p *Protocol) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaler interface.
 func (p Protocol) MarshalJSON() ([]byte, error) {
 	return json.Marshal(p.String())
-}
-
-// LocalizationMethod represents an enabled localization method for a cachegroup. The string values of this type should match the Traffic Ops values.
-type LocalizationMethod string
-
-const (
-	LocalizationMethodCZ      = LocalizationMethod("CZ")
-	LocalizationMethodDeepCZ  = LocalizationMethod("DEEP_CZ")
-	LocalizationMethodGeo     = LocalizationMethod("GEO")
-	LocalizationMethodInvalid = LocalizationMethod("INVALID")
-)
-
-// String returns a string representation of this localization method
-func (m LocalizationMethod) String() string {
-	switch m {
-	case LocalizationMethodCZ:
-		return string(m)
-	case LocalizationMethodDeepCZ:
-		return string(m)
-	case LocalizationMethodGeo:
-		return string(m)
-	default:
-		return "INVALID"
-	}
-}
-
-func LocalizationMethodFromString(s string) LocalizationMethod {
-	switch strings.ToLower(s) {
-	case "cz":
-		return LocalizationMethodCZ
-	case "deep_cz":
-		return LocalizationMethodDeepCZ
-	case "geo":
-		return LocalizationMethodGeo
-	default:
-		return LocalizationMethodInvalid
-	}
-}
-
-func (m *LocalizationMethod) UnmarshalJSON(data []byte) error {
-	if string(data) == "null" {
-		return errors.New("LocalizationMethod cannot be null")
-	}
-	s, err := strconv.Unquote(string(data))
-	if err != nil {
-		return errors.New(string(data) + " JSON not quoted")
-	}
-	*m = LocalizationMethodFromString(s)
-	if *m == LocalizationMethodInvalid {
-		return errors.New(s + " is not a LocalizationMethod")
-	}
-	return nil
-}
-
-func (m LocalizationMethod) MarshalJSON() ([]byte, error) {
-	return json.Marshal(m.String())
-}
-
-func (m *LocalizationMethod) Scan(value interface{}) error {
-	if value == nil {
-		return errors.New("LocalizationMethod cannot be null")
-	}
-	sv, err := driver.String.ConvertValue(value)
-	if err != nil {
-		return errors.New("failed to scan LocalizationMethod: " + err.Error())
-	}
-
-	switch v := sv.(type) {
-	case []byte:
-		*m = LocalizationMethodFromString(string(v))
-		if *m == LocalizationMethodInvalid {
-			return errors.New(string(v) + " is not a valid LocalizationMethod")
-		}
-		return nil
-	default:
-		return fmt.Errorf("failed to scan LocalizationMethod, unsupported input type: %T", value)
-	}
 }
 
 // DeepCachingType represents a Delivery Service's deep caching type. The string values of this type should match the Traffic Ops values.
